server_trader/requirements.txt
longport>=1.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
schedule>=1.2.0
requests>=2.31.0
pytz>=2023.3
